refactor (utils): move utility files
