+++ src/api/v2/ontology/ontologies-endpoint.ts
@@ -3,7 +3,7 @@ import { AjaxResponse } from "rxjs/ajax";
 import { catchError, map, mergeMap } from "rxjs/operators";
 import { ApiResponseError } from "../../..";
 import { ReadOntology } from "../../../models/v2/ontologies/read-ontology";
-import { OntologyConversionUtils } from "../../../models/v2/OntologyConversionUtil";
+import { OntologyConversionUtils } from "../../../models/v2/ontologies/OntologyConversionUtil";
 import { Endpoint } from "../../endpoint";
 
 declare let require: any; // http://stackoverflow.com/questions/34730010/angular2-5-minute-install-bug-require-is-not-defined
