+++ src/api/v2/resource/resources-endpoint.ts
@@ -3,7 +3,7 @@ import { AjaxResponse } from "rxjs/ajax";
 import { catchError, map, mergeMap } from "rxjs/operators";
 import { ApiResponseError, OntologyCache } from "../../..";
 import { ReadResource } from "../../../models/v2/resources/read-resource";
-import { ResourcesConversionUtils } from "../../../models/v2/ResourcesConversionUtils";
+import { ResourcesConversionUtils } from "../../../models/v2/resources/ResourcesConversionUtils";
 import { Endpoint } from "../../endpoint";
 
 declare let require: any; // http://stackoverflow.com/questions/34730010/angular2-5-minute-install-bug-require-is-not-defined
